+++ Swiften/LinkLocal/DNSSD/Bonjour/BonjourBrowseQuery.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010 Remko Tronçon
+ * Copyright (c) 2010-2013 Remko Tronçon
  * Licensed under the GNU General Public License v3.
  * See Documentation/Licenses/GPLv3.txt for more information.
  */
@@ -9,6 +9,7 @@
 #include <Swiften/LinkLocal/DNSSD/Bonjour/BonjourQuery.h>
 #include <Swiften/LinkLocal/DNSSD/DNSSDBrowseQuery.h>
 #include <Swiften/EventLoop/EventLoop.h>
+#include <boost/numeric/conversion/cast.hpp>
 
 namespace Swift {
 	class BonjourQuerier;
@@ -48,7 +49,7 @@ namespace Swift {
 				}
 				else {
 					//std::cout << "Discovered service: name:" << name << " domain:" << domain << " type: " << type << std::endl;
-					DNSSDServiceID service(name, domain, type, interfaceIndex);
+					DNSSDServiceID service(name, domain, type, boost::numeric_cast<int>(interfaceIndex));
 					if (flags & kDNSServiceFlagsAdd) {
 						eventLoop->postEvent(boost::bind(boost::ref(onServiceAdded), service), shared_from_this());
 					}
