+++ Swiften/LinkLocal/DNSSD/Bonjour/BonjourResolveHostnameQuery.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010 Remko Tronçon
+ * Copyright (c) 2010-2013 Remko Tronçon
  * Licensed under the GNU General Public License v3.
  * See Documentation/Licenses/GPLv3.txt for more information.
  */
@@ -13,6 +13,7 @@
 #include <Swiften/LinkLocal/DNSSD/DNSSDResolveHostnameQuery.h>
 #include <Swiften/EventLoop/EventLoop.h>
 #include <Swiften/Network/HostAddress.h>
+#include <boost/numeric/conversion/cast.hpp>
 
 #include <netinet/in.h>
 
@@ -23,7 +24,7 @@ namespace Swift {
 		public: 
 			BonjourResolveHostnameQuery(const std::string& hostname, int interfaceIndex, boost::shared_ptr<BonjourQuerier> querier, EventLoop* eventLoop) : BonjourQuery(querier, eventLoop) {
 				DNSServiceErrorType result = DNSServiceGetAddrInfo(
-						&sdRef, 0, interfaceIndex, kDNSServiceProtocol_IPv4, 
+						&sdRef, 0, boost::numeric_cast<unsigned int>(interfaceIndex), kDNSServiceProtocol_IPv4, 
 						hostname.c_str(), 
 						&BonjourResolveHostnameQuery::handleHostnameResolvedStatic, this);
 				if (result != kDNSServiceErr_NoError) {
